Fix DataDownloads script errors and improve timeout configuration
This commit resolves all failing DataDownloads scripts and enhances the timeout system:

## Script Fixes:
- **C_CompustatQuarterly.py**: Fixed Polars data type error by adding explicit Float64 casting
- **ZB_PIN.py**: Fixed Dropbox download timeout by updating to correct redirected URL and improving download logic
- **ZC_GovernanceIndex.py**: Fixed UnboundLocalError by correcting variable name from 'data' to 'final_data'
- **ZI_PatentCitations.py**: Fixed IndentationError by properly structuring the placeholder logic
- **ZL_CRSPOPTIONMETRICS.py**: Fixed IndentationError and missing f-string formatting

## Timeout Configuration Enhancement:
- **config.py**: Added configurable `SCRIPT_TIMEOUT_MINUTES = 10` (changed from hardcoded 30 seconds)
- **01_DownloadData.py**: Updated to use configurable timeout from config.py with minute-based units
- Timeout messages now display minutes instead of seconds for better readability
- Default 10-minute timeout handles large downloads like ZB_PIN.py (273MB file)

## Results:
- All DataDownloads scripts now complete with return code 0
- ZB_PIN.py successfully downloads and processes actual PIN data (21,206 yearly → 254,472 monthly records)
- Configurable timeout system allows easy adjustment for different environments (0.5 min for testing, 10 min for production)

+++ Signals/pyCode/DataDownloads/C_CompustatQuarterly.py
@@ -155,9 +155,9 @@ for var in ytd_vars:
         # Create quarterly version: Q1 gets full value, Q2-Q4 get difference from previous
         ytd_exprs.append(
             pl.when(pl.col('fqtr') == 1)
-            .then(pl.col(var))
+            .then(pl.col(var).cast(pl.Float64))
             .otherwise(
-                pl.col(var) - pl.col(var).shift(1).over(['gvkey', 'fyearq'])
+                pl.col(var).cast(pl.Float64) - pl.col(var).cast(pl.Float64).shift(1).over(['gvkey', 'fyearq'])
             )
             .alias(var_q)
         )
